Adding support for armv6l daemon builds
With the addition of another xCPUARCH parameter check for armv6l,
we enable the daemon/headless wallet build in spirit. In truth,
it was already possible given that xCPUARCH is passed in as a
define, but it makes building less confusing for people that
don't know what they're doing.

The relevant documentation has been updated for armv6l as well.