Feature: check for updated binaries
A file on a remote server (version.txt for example) that contains
the newest wallet information is retrieved, then the string in
that file is parsed and compared with the internal wallet version,
which is defined in src/version.h. If any integer in the remote
version is larger than the corresponding internal integer, a
clickable label is displayed in the GUI overview page, prompting
the user to download the new wallet version. This check occurs
once per day, as commited.

---
Notes:

The remote version.txt must contain only a string of the form:
MAJOR_VERSION.MINOR_VERSION.REVISION.BUILD (e.g. 1.2.1.2) with no
leading or trailing spaces.

The URLs for the remote version.txt file and the suggested
download location URL are defined in src/qt/updatecheck.h. The
Update time interval is also defined in src/qt/updatecheck.h.

The UpdateCheck class is used in the OverviewPage class
(src/qt/overviewpage.cpp) along with a little extra logic to
achieve this functionality.